requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
websockets
dnspython
pydantic
python-multipart
starlette
requests
